requirements.txt
PySide6==6.6.1
requests==2.31.0
pandas==2.1.4
pyyaml==6.0.1
openai==1.12.0
tqdm==4.66.1 
openpyxl